Speed up CPU demeaning ~20%: Int32 refs and fast LSMR 2-norm (#81)
Two CPU-side optimizations on top of #80, with no change to results.

- FixedEffect: store group refs as Int32 when ngroups <= typemax(Int32),
  halving the dominant memory stream read by the scatter/gather kernels on
  every solver iteration. Every backend (CPU/GPU) and solve_coefficients!
  benefits; ngroups beyond Int32 keeps the original integer type.

- lsmr!: replace LinearAlgebra.norm's overflow-safe (~8x slower) path with a
  SIMD sum-of-squares at the four hot norm call sites, accumulated in Float64
  for accuracy. Restricted to concrete CPU Vector{Float32,Float64}; GPU
  arrays and FixedEffectCoefficients keep the generic norm.

Bumps version to 3.3.0.

+++ test/types.jl
@@ -9,14 +9,14 @@ import Base.==
     @test sprint(show, fe1) == "Fixed Effects"
     @test sprint(show, MIME("text/plain"), fe1) == """
         Fixed Effects:
-          refs (10-element Vector{Int64}):
+          refs (10-element Vector{Int32}):
             [1, 2, 3, 4, 5, ... ]
           interaction (UnitWeights):
             none"""
     fe2 = FixedEffect(1:10, interaction=fill(1.23456789, 10))
     @test sprint(show, MIME("text/plain"), fe2) == """
         Fixed Effects:
-          refs (10-element Vector{Int64}):
+          refs (10-element Vector{Int32}):
             [1, 2, 3, 4, 5, ... ]
           interaction (10-element Vector{Float64}):
             [1.23457, 1.23457, 1.23457, 1.23457, 1.23457, ... ]"""
